+++ src/ARPES.jl
@@ -13,6 +13,7 @@ export merge_consensus
 export kx, ky, kz, phi, psi, eV, detector_ch, ch2, delay, spin
 export CPS, Counts
 export ARPESData
+include("./stitch.jl")
 include("./pumpprobe.jl")
 include("./transform.jl")